feat : 컴포넌트 공동 사용 hook분리 및 적용
