+++ src/components/DynamicBtn/CheckBoxBtn.tsx
@@ -1,5 +1,5 @@
-import { useState } from "react";
 import type { BaseBtnProps, IconTextStatus } from "../../types/DynamicBtn";
+import useDynamicStatus from "../../hooks/useDynamicStatus";
 
 export default function CheckBoxBtn({
   children,
@@ -8,18 +8,11 @@ export default function CheckBoxBtn({
   type,
 }: BaseBtnProps) {
   //상태 관리
-  const [clicked, setClicked] = useState(false);
-  const [isPressing, setIsPressing] = useState(false);
 
-  const getStatus = (): IconTextStatus => {
-    if (disabled) return "disabled";
-    if (clicked && isPressing) return "CLpressing";
-    if (isPressing) return "pressing";
-    if (clicked) return "clicked";
-    return "default";
-  };
-
-  const status: IconTextStatus = getStatus();
+  const { status, onMouseDown, onMouseLeave, onMouseUp } = useDynamicStatus(
+    disabled,
+    true,
+  );
 
   const statusMap: Record<IconTextStatus, { bg: string; icon: string }> = {
     clicked: {
@@ -49,17 +42,13 @@ export default function CheckBoxBtn({
   return (
     <button
       className={`inline-flex pr-2 py-1 pl-[0.375rem] gap-2 rounded-lg items-center body-rg-500  ${bg} ${disabled ? "cursor-not-allowed" : "cursor-pointer"}`}
-      onMouseDown={() => !disabled && setIsPressing(true)}
-      onMouseUp={() => {
-        if (disabled) return;
-        setIsPressing(false);
-        setClicked(prev => !prev); // 토글
-      }}
-      onMouseLeave={() => !disabled && setIsPressing(false)}
+      onMouseDown={onMouseDown}
+      onMouseUp={onMouseUp}
+      onMouseLeave={onMouseLeave}
       onClick={onClick}
       type={type ? type : "button"}
     >
-      <img src={icon} alt="" className="pt-[0.125rem] size-4" />
+      <img src={icon} alt="" className=" size-4" />
       {children}
     </button>
   );
