+++ src/components/DynamicBtn/DropDownBtn.tsx
@@ -1,5 +1,5 @@
-import { useState } from "react";
 import type { BaseBtnProps, TextIconStatus } from "../../types/DynamicBtn";
+import useDynamicStatus from "../../hooks/useDynamicStatus";
 
 export default function DropDownBtn({
   children,
@@ -8,17 +8,11 @@ export default function DropDownBtn({
   type,
 }: BaseBtnProps) {
   //상태 관리
-  const [clicked, setClicked] = useState(false);
-  const [isPressing, setIsPressing] = useState(false);
 
-  const getStatus = (): TextIconStatus => {
-    if (disabled) return "disabled";
-    if (isPressing) return "pressing";
-    if (clicked) return "clicked";
-    return "default";
-  };
-
-  const status: TextIconStatus = getStatus();
+  const { status, onMouseDown, onMouseLeave, onMouseUp } = useDynamicStatus(
+    disabled,
+    false,
+  );
 
   const statusMap: Record<TextIconStatus, { bg?: string; icon: string }> = {
     clicked: {
@@ -35,19 +29,14 @@ export default function DropDownBtn({
       icon: "/src/assets/icons/arrow_down.svg",
     },
   };
-
-  const { bg, icon } = statusMap[status];
+  const { bg, icon } = statusMap[status as TextIconStatus];
 
   return (
     <button
       className={`inline-flex pr-[0.625rem] py-1 pl-4 gap-1 rounded-lg items-center header-h4  ${bg} ${disabled ? "cursor-not-allowed" : "cursor-pointer"}`}
-      onMouseDown={() => !disabled && setIsPressing(true)}
-      onMouseUp={() => {
-        if (disabled) return;
-        setIsPressing(false);
-        setClicked(prev => !prev); // 토글
-      }}
-      onMouseLeave={() => !disabled && setIsPressing(false)}
+      onMouseDown={onMouseDown}
+      onMouseUp={onMouseUp}
+      onMouseLeave={onMouseLeave}
       onClick={onClick}
       type={type ? type : "button"}
     >
