+++ src/components/DynamicBtn/TabBtn.tsx
@@ -1,5 +1,5 @@
-import { useState } from "react";
 import type { BaseBtnProps, TextIconStatus } from "../../types/DynamicBtn";
+import useDynamicStatus from "../../hooks/useDynamicStatus";
 
 export default function TabBtn({
   children,
@@ -7,18 +7,10 @@ export default function TabBtn({
   onClick,
   type,
 }: BaseBtnProps) {
-  //click 했을때 span생김
-  //pressing헸을때 bg-gy-100
-  //disable 했을때 text-gy-400
-  const [clicked, setClicked] = useState(false);
-  const [isPressing, setIsPressing] = useState(false);
-  const getStatus = (): TextIconStatus => {
-    if (disabled) return "disabled";
-    if (isPressing) return "pressing";
-    if (clicked) return "clicked";
-    return "default";
-  };
-  const status: TextIconStatus = getStatus();
+  const { status, onMouseDown, onMouseLeave, onMouseUp } = useDynamicStatus(
+    disabled,
+    false,
+  );
 
   const statusMap: Record<TextIconStatus, { bg?: string; span?: string }> = {
     clicked: {
@@ -35,18 +27,14 @@ export default function TabBtn({
     },
   };
 
-  const { bg, span } = statusMap[status];
+  const { bg, span } = statusMap[status as TextIconStatus];
 
   return (
     <button
       className={`header-h5 inline-flex  flex-col justify-center items-start  rounded-lg relative py-2 gap-1 px-3 ${bg}`}
-      onMouseDown={() => !disabled && setIsPressing(true)}
-      onMouseUp={() => {
-        if (disabled) return;
-        setIsPressing(false);
-        setClicked(prev => !prev); // 토글
-      }}
-      onMouseLeave={() => !disabled && setIsPressing(false)}
+      onMouseDown={onMouseDown}
+      onMouseUp={onMouseUp}
+      onMouseLeave={onMouseLeave}
       onClick={onClick}
       type={type ? type : "button"}
     >
