+++ src/components/DynamicBtn/TagBtn.tsx
@@ -1,4 +1,4 @@
-import { useState } from "react";
+import useDynamicStatus from "../../hooks/useDynamicStatus";
 import type { BaseBtnProps, IconTextStatus } from "../../types/DynamicBtn";
 
 export default function TagBtn({
@@ -7,19 +7,10 @@ export default function TagBtn({
   onClick,
   type,
 }: BaseBtnProps) {
-  //상태 관리
-  const [clicked, setClicked] = useState(false);
-  const [isPressing, setIsPressing] = useState(false);
-
-  const getStatus = (): IconTextStatus => {
-    if (disabled) return "disabled";
-    if (clicked && isPressing) return "CLpressing";
-    if (isPressing) return "pressing";
-    if (clicked) return "clicked";
-    return "default";
-  };
-
-  const status: IconTextStatus = getStatus();
+  const { status, onMouseDown, onMouseLeave, onMouseUp } = useDynamicStatus(
+    disabled,
+    true,
+  );
 
   const statusMap: Record<IconTextStatus, { bg: string; icon: string }> = {
     clicked: {
@@ -49,13 +40,9 @@ export default function TagBtn({
   return (
     <button
       className={`inline-flex pr-[0.625rem] py-2 pl-3 gap-1 rounded-2xl items-center body-rg-500 shadow-[0px_0px_12px_0px_rgba(18,18,18,0.04)] ${bg} ${disabled ? "cursor-not-allowed" : "cursor-pointer"}`}
-      onMouseDown={() => !disabled && setIsPressing(true)}
-      onMouseUp={() => {
-        if (disabled) return;
-        setIsPressing(false);
-        setClicked(prev => !prev); // 토글
-      }}
-      onMouseLeave={() => !disabled && setIsPressing(false)}
+      onMouseDown={onMouseDown}
+      onMouseUp={onMouseUp}
+      onMouseLeave={onMouseLeave}
       onClick={onClick}
       type={type ? type : "button"}
     >
