+++ src/components/DynamicBtn/TextIconBtnS.tsx
@@ -1,5 +1,5 @@
-import { useState } from "react";
 import type { BaseBtnProps, TextIconStatus } from "../../types/DynamicBtn";
+import useDynamicStatus from "../../hooks/useDynamicStatus";
 
 export default function TextIconBtnS({
   children,
@@ -7,19 +7,10 @@ export default function TextIconBtnS({
   onClick,
   type,
 }: BaseBtnProps) {
-  //상태 관리
-  const [clicked, setClicked] = useState(false);
-  const [isPressing, setIsPressing] = useState(false);
-
-  const getStatus = (): TextIconStatus => {
-    if (disabled) return "disabled";
-    if (isPressing) return "pressing";
-    if (clicked) return "clicked";
-    return "default";
-  };
-
-  const status: TextIconStatus = getStatus();
-
+  const { status, onMouseDown, onMouseLeave, onMouseUp } = useDynamicStatus(
+    disabled,
+    false,
+  );
   const statusMap: Record<TextIconStatus, { bg?: string; icon: string }> = {
     clicked: {
       icon: "/src/assets/icons/arrow_right.svg",
@@ -36,18 +27,14 @@ export default function TextIconBtnS({
     },
   };
 
-  const { bg, icon } = statusMap[status];
+  const { bg, icon } = statusMap[status as TextIconStatus];
 
   return (
     <button
       className={`inline-flex pl-2 py-1 pr-1 gap-1 rounded-lg items-center  body-rg-400 ${bg} ${disabled ? "cursor-not-allowed" : "cursor-pointer"}`}
-      onMouseDown={() => !disabled && setIsPressing(true)}
-      onMouseUp={() => {
-        if (disabled) return;
-        setIsPressing(false);
-        setClicked(prev => !prev); // 토글
-      }}
-      onMouseLeave={() => !disabled && setIsPressing(false)}
+      onMouseDown={onMouseDown}
+      onMouseUp={onMouseUp}
+      onMouseLeave={onMouseLeave}
       onClick={onClick}
       type={type ? type : "button"}
     >
