--- src/constants/variantStyle.ts
@@ -1,65 +1,0 @@
-// //variant별 기본 style정리
-
-// import type { Variant, VariantStyleConfig } from "../types/DynamicBtn";
-
-// export const variantStyleMap: Record<Variant, VariantStyleConfig> = {
-//   keyword: {
-//     default: {
-//       size: "md", //ok
-//       bg: "white", //ok
-//       borderTheme: "gray", //ok
-//       shadow: "ds50", //ok
-//       rounded: "sm", //ok
-//     },
-//     clicked: {
-//       size: "md", //ok
-//       bg: "gray", //ok
-//       borderTheme: "green", //ok
-//       shadow: "ds50", //ok
-//       rounded: "sm", //ok
-//     },
-//     pressing: {
-//       size: "md", //ok
-//       bg: "gray2", //ok
-//       borderTheme: "green", //ok
-//       shadow: "ds50", //ok
-//       rounded: "sm", //ok
-//     },
-//   },
-//   checkbox: {
-//     size: "sm",
-//     bg: "none",
-//     borderTheme: "gray",
-//     shadow: "none",
-//   },
-//   dropdown: {
-//     size: "lg",
-//     bg: "white",
-//     borderTheme: "none",
-//     shadow: "none",
-//   },
-//   filter: {
-//     size: "sm",
-//     bg: "white",
-//     borderTheme: "none",
-//     shadow: "none",
-//   },
-//   tab: {
-//     size: "xs",
-//     bg: "none",
-//     borderTheme: "none",
-//     shadow: "none",
-//   },
-//   plain: {
-//     size: "sm",
-//     bg: "none",
-//     borderTheme: "none",
-//     shadow: "none",
-//   },
-//   gy100: {
-//     size: "sm",
-//     bg: "gray",
-//     borderTheme: "none",
-//     shadow: "ds100",
-//   },
-// };
