+++ src/types/DynamicBtn.ts
@@ -1,11 +1,6 @@
-export type IconTextStatus =
-  | "clicked"
-  | "CLpressing"
-  | "pressing"
-  | "default"
-  | "disabled";
-
 export type TextIconStatus = "clicked" | "pressing" | "default" | "disabled";
+
+export type IconTextStatus = TextIconStatus | "CLpressing";
 
 export type BaseBtnProps = {
   children: React.ReactNode;
